requirements.txt
torch>=2.1
torchvision>=0.16
opencv-python
tqdm
pyyaml
numpy
matplotlib
scipy